Auto-pulled modifications into their own assembly
Inlcuded in otapi.patcher.engine is a patch.sh script which takes all
the modification bases out of the engine and scaffolds a new DNX
project for each of them, updates the project.json files and makes
them all buildable.  All of the generated projects build and are
added to the solution.

Requires node.js and the json-cli npm package installed to work.