WIP - this will become the base of v3
GPL is now implemented, this branch is to mainly hold my code until the official branch is ready. I have decided on MonoMod, with some extra OTAPI functionality as it doesnt fit all and using MonoMod rules can be combersome, but the power of patching and runtime hooks are awesome. I think this may be a good foundation