Stack counting + GetEntityEdgeTile index out of range port + WIP SendData lock removals
The stack counter is likely the most important part here. It's needed for rewriting of locks, so i can determine the instructions on stack and then being able to remove the unneeded code to remove the try/finally blocks fully.
Unlike old takes on this stack counter where it only kept track of the current instruction set from top to bottom, usually incorrectly, this one is able to traverse up and down by computing the full stack ahead of time.
It seems to work for my use case atm, however i havent really considered branching/arcs yet so it's possible that i havent computed the stack correctly in all scenarios yet.