1.4.2.3 + .net5 retargeting + examples
.net5 retargeting is mostly the focus in this.

In order to be able to merge in .net5 libraries via patching they would need to target the same frameworks. terraria is built on mscorlib and .net4 whereas patches and mods are via System.Runtime/CoreLib on .net5. CoreLibRelinker was made in order to scan over an assembly as a post process step and it will translate mscorlib and Private CoreLib references to their respective facade libraries.

this implementation is no where near complete, but it seems to be on a promising track as it appears to work on windows and osx, with at least one player, all hoisted on .net5 with .net5 patches running directly in the terraria binary, inc monomod events at runtime.

additionally in this commit i've added a couple example runtime plugins, one as a script the other as a compiled dll. both effectively can do the same thing, just different ways of creating a plugin.

however, this will not likely be used by the likes of TShock/Orion etc. this plugin system is mainly just for extensibility and for local projects/extensions. i am also aware of other projects that would be possible to directly integrate into a universal plugin (ie ones built for otapi, but maybe otapi also runs tshock and orion so this plugin can run on either tshock or orion, e.g. i've seen tiled as a great candidate for this if it's built for otapi3)