Client now runs on net5, and can install the self-contained framework.
ALL SHIMS ARE NOW SCRIPTS! This is why this is such a big commit.

Further more, c# scripts are now loaded from csharp/plugins/[toplevel|modules|patches]

Installer was also rewritten in order to procude a published dotnet output, and then install it to the target.

NOTE: windows install targets may be included but not yet working. haven't had much time on a windows machine yet to work this out correctly.